requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Jira and Confluence integration
atlassian-python-api==3.41.4
jira==3.5.1

# Data handling
pandas==2.2.0
pyyaml==6.0.1
pydantic==2.6.1

# Scheduling
APScheduler==3.10.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0

# Logging and monitoring
python-json-logger==2.0.7

# Development tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Type hinting
typing-extensions==4.9.0

# HTML parsing (for Confluence tables)
lxml==5.1.0
html5lib==1.1
beautifulsoup4==4.12.3

# Environment variables
python-dotenv==1.0.1